requirements.txt
streamlit
google-generativeai
pandas
streamlit-webrtc
torch
transformers
Pillow
timm
streamlit-modal